Move entry scripts to package
