+++ scripts/coordinate_finder.py
