+++ scripts/getcontact_phone_checker.py
@@ -1,4 +1,4 @@
-from phone_checker_cli import main
+from .phone_checker_cli import main
 
 if __name__ == "__main__":
     exit(main(["getcontact"]))
