+++ scripts/phone_checker_cli.py
